+++ index.js
@@ -1,6 +1,5 @@
-import { navBarListener } from './module/navbar.js';
+import { navBarListener, timeFunction } from './module/navbar.js';
 import * as bookController from './module/bookOperations.js';
-import { timeFunction } from './module/time.js';
 
 navBarListener();
 setInterval(() => timeFunction(), 1000);
