+++ module/bookLibrary.js
@@ -1,4 +1,4 @@
-export class BooksLibrary {
+export default class BooksLibrary {
   constructor() {
     this.bookArray = (JSON.parse(localStorage.getItem('booksDb')) !== null) ? (JSON.parse(localStorage.getItem('booksDb'))) : ([]);
   }
